Fix for upsert behavior (#193)
* Fix for upsert behavior

* psalm fix

* disable rector

+++ tests/Support/Fixture/pgsql.sql
@@ -389,7 +389,7 @@ CREATE TABLE "T_constraints_6"
 CREATE TABLE "T_upsert"
 (
     "id" SERIAL NOT NULL PRIMARY KEY,
-    "ts" INT NULL,
+    "ts" BIGINT NULL,
     "email" VARCHAR(128) NOT NULL UNIQUE,
     "recovery_email" VARCHAR(128) NULL,
     "address" TEXT NULL,
